zynvaro-app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
scikit-learn==1.5.2
numpy==1.26.4
httpx==0.27.2
python-dotenv==1.0.1
apscheduler==3.10.4
anthropic==0.40.0
pytest==9.0.2
pytest-asyncio==0.23.8
